Finished REST API for users, loans and items

+++ backend/src/main/java/com/student_loan/controller/UserController.java
@@ -6,11 +6,12 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import com.student_loan.model.User;
 import com.student_loan.service.UserService;
-import com.student_loan.dtos.UserDTO;
+import com.student_loan.dtos.UserRecord;
 import com.student_loan.dtos.CredentialsDTO;
+import com.student_loan.dtos.RegistrationRecord;
 
+import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 @RestController
 @RequestMapping("/users")
@@ -19,8 +20,12 @@ public class UserController {
     private UserService userService;
 
     @GetMapping
-    public List<User> getAllUsers() {
-        return userService.getAllUsers();
+    public ResponseEntity<List<User>> getAllUsers(@RequestParam("token") String token) {
+    	 User user = userService.getUserByToken(token);
+        if (user == null || user.getAdmin()==false) {
+        	return new ResponseEntity<>(new ArrayList<>(),HttpStatus.UNAUTHORIZED);
+        }
+    	return new ResponseEntity<>(userService.getAllUsers(), HttpStatus.OK);
     }
 
     @PostMapping("/login")
@@ -43,14 +48,24 @@ public class UserController {
 	}
     
     @GetMapping("/{id}")
-    public Optional<User> getUserById(@PathVariable Long id) {
-        return userService.getUserById(id);
+    public ResponseEntity<User> getUserById(@PathVariable Long id, @RequestParam("token") String token) {
+        
+    	User user = userService.getUserByToken(token);
+        if (user == null || user.getAdmin()==false && user.getId()!=id) {
+        	   return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+    	return new ResponseEntity<>(userService.getUserById(id).get(), HttpStatus.OK);
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<User> updateUser(@PathVariable Long id, @RequestBody User userData) {
-        try {
-            User updatedUser = userService.updateUser(id, userData);
+    public ResponseEntity<User> updateUser(@PathVariable Long id, @RequestBody UserRecord userData, @RequestParam("token") String token) {
+        User user = userService.getUserByToken(token);
+		if (user == null || user.getAdmin() == false && user.getId() != id) {
+			return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+		}
+    	
+    	try {
+            User updatedUser = userService.updateUser(id, userRecordToUser(userData));
             return new ResponseEntity<>(updatedUser, HttpStatus.OK);
         } catch (RuntimeException e) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
@@ -60,8 +75,8 @@ public class UserController {
 
     //Register 
     @PostMapping("/register")
-    public ResponseEntity<String> register(@RequestBody UserDTO userDTO) {
-        if (userService.register(dtoToUser(userDTO))) {
+    public ResponseEntity<String> register(@RequestBody RegistrationRecord userDTO) {
+        if (userService.register(registerRecordToUser(userDTO))) {
             return ResponseEntity.ok("User registered correctly");
         } else {
             return ResponseEntity.badRequest().body("The user already exists");
@@ -69,21 +84,44 @@ public class UserController {
     }
 
     @DeleteMapping("/{id}")
-    public void deleteUser(@PathVariable Long id) {
-        userService.deleteUser(id);
+    public ResponseEntity<String> deleteUser(@PathVariable Long id, @RequestParam("token") String token) {
+        User user = userService.getUserByToken(token);
+		if (user == null || user.getAdmin() == false && user.getId() != id) {
+			return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+		}
+    	userService.deleteUser(id);
+
+    	return new ResponseEntity<>(HttpStatus.OK);
+    }
+    private User registerRecordToUser(RegistrationRecord data) {
+    	        User user = new User();
+    	        user.setName(data.name() + " " + data.lastName() );
+    	        user.setEmail(data.email());
+    	        user.setPassword(data.password());
+    	        user.setTelephoneNumber("");
+    	        user.setAddress("");
+    			user.setDegreeType(User.DegreeType.UNIVERSITY_DEGREE);
+    			user.setDegreeYear(0);
+    			user.setPenalties(0);
+    			user.setAverageRating(0.0);
+    			user.setAdmin(false);
+    			return user;
+    			
+    	                
     }
     
-	private User dtoToUser(UserDTO userDTO) {
+	private User userRecordToUser(UserRecord userDTO) {
 		User user = new User();
-		user.setName(userDTO.getName()+" "+userDTO.getLastName());
-		user.setEmail(userDTO.getEmail());
-		user.setPassword(userDTO.getPassword());
-		user.setTelephoneNumber("");
-		user.setAddress("");
+		user.setName(userDTO.name()+" "+userDTO.lastName());
+		user.setEmail(userDTO.email());
+		user.setPassword(userDTO.password());
+		user.setTelephoneNumber(userDTO.telephoneNumber());
+		user.setAddress(userDTO.address());
 		user.setDegreeType(User.DegreeType.UNIVERSITY_DEGREE);
-		user.setDegreeYear(0);
-		user.setPenalties(0);
-		user.setAverageRating(0.0);
+		user.setDegreeYear(userDTO.degreeYear());
+		user.setPenalties(userDTO.penalties());
+		user.setAverageRating(userDTO.averageRating());
+		user.setAdmin(userDTO.admin());
 		
 		return user;
 	}
